first crack at paging between UIViews
This is way more difficult than it should be. Giving up for now and transitioning to tab-based switching — which, come to think of it, may be more efficient. Committing for posterity.

Should one come back around to swipe-based pagination, consider implementing the data model as an array of NSDictionaries instead of individual variables. This should allow index querying based on view, as in the standard example from Apple.